# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  description: API do app MoveUp. Erros em application/problem+json (RFC 9457) com
    `code` estável.
  title: MoveUp API
  version: v1
servers:
- url: /
security:
- bearerAuth: []
...
paths:
...
  /v1/alert-settings:
    get:
      operationId: getAlertSettings
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlertSetting"
          description: Uma por tipo
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Conta não é de profissional (`forbidden`)
      summary: Configurações de alerta (com padrões)
      tags:
      - alerts
    put:
      operationId: updateAlertSettings
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertSettingsUpdate"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlertSetting"
          description: Todas as configurações depois da mudança
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Conta não é de profissional (`forbidden`)
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "`threshold-invalid`, `alert-type-invalid`"
      summary: Altera configurações de alerta
      tags:
      - alerts
  /v1/alerts:
    get:
      description: "status: open (padrão; inclui adiados vencidos), snoozed ou resolved."
      operationId: listAlerts
      parameters:
      - in: query
        name: status
        required: false
        schema:
          type: string
      - in: query
        name: before
        required: false
        schema:
          type: string
          format: uuid
      - in: query
        name: limit
        required: false
        schema:
          type: integer
          format: int32
          default: 30
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertPage"
          description: Uma página de alertas
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Conta não é de profissional (`forbidden`)
      summary: Alertas do profissional (mais novos primeiro)
      tags:
      - alerts
  /v1/alerts/{alertId}/resolve:
    post:
      operationId: resolveAlert
      parameters:
      - in: path
        name: alertId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Resolvido (repetir não muda nada)
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Conta não é de profissional (`forbidden`)
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: '`resource-not-found`'
      summary: Marca o alerta como resolvido
      tags:
      - alerts
  /v1/alerts/{alertId}/snooze:
    post:
      operationId: snoozeAlert
      parameters:
      - in: path
        name: alertId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertSnooze"
        required: true
      responses:
        "204":
          description: Adiado
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Conta não é de profissional (`forbidden`)
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "`snooze-invalid`, `alert-resolved`"
      summary: Adia o alerta por alguns dias
      tags:
      - alerts
...
  /v1/push-devices:
    put:
      operationId: registerPushDevice
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushDevice"
        required: true
      responses:
        "204":
          description: Registrado (repetir atualiza)
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Conta não é de profissional (`forbidden`)
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "`push-token-invalid`, `platform-invalid`"
      summary: Registra o aparelho para push
      tags:
      - alerts
  /v1/push-devices/remove:
    post:
      operationId: removePushDevice
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushDeviceRemoval"
        required: true
      responses:
        "204":
          description: Removido (ou já não existia)
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Conta não é de profissional (`forbidden`)
      summary: Tira o aparelho do push (ao sair da conta)
      tags:
      - alerts
...
components:
  schemas:
...
    AlertItem:
      type: object
      properties:
        clientId:
          type: string
          format: uuid
        clientName:
          type: string
        createdAt:
          type: string
          format: date-time
        facts:
          type: object
          additionalProperties:
            type: integer
            format: int32
          description: "Só números: days, percent, planned, done, effort, count"
        id:
          type: string
          format: uuid
        linkId:
          type: string
          format: uuid
          description: Vínculo atual do aluno com o profissional
        severity:
          type: string
          enum:
          - info
          - warning
          - urgent
        snoozedUntil:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - open
          - snoozed
          - resolved
        type:
          type: string
          enum:
          - pain_reported
          - high_effort
          - new_feedback
          - session_edited
          - inactive
          - low_adherence
          - clearance_pending
      required:
      - clientId
      - createdAt
      - facts
      - id
      - severity
      - status
      - type
    AlertPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AlertItem"
        next:
          type: string
          format: uuid
          description: Cursor da próxima página (ausente no fim)
        openCount:
          type: integer
          format: int32
          description: Abertos no total (badge)
      required:
      - items
      - openCount
    AlertSetting:
      type: object
      properties:
        enabled:
          type: boolean
        push:
          type: boolean
        threshold:
          type: integer
          format: int32
          description: "inactive: dias sem treinar (2–60); high_effort: esforço (5–\
            10); low_adherence: % mínima (10–100). Ausente nos tipos sem limite"
        type:
          type: string
          enum:
          - pain_reported
          - high_effort
          - new_feedback
          - session_edited
          - inactive
          - low_adherence
          - clearance_pending
          minLength: 1
      required:
      - enabled
      - push
      - type
    AlertSettingsUpdate:
      type: object
      properties:
        settings:
          type: array
          items:
            $ref: "#/components/schemas/AlertSetting"
          maxItems: 20
          minItems: 0
      required:
      - settings
    AlertSnooze:
      type: object
      properties:
        days:
          type: integer
          format: int32
          description: 1 a 30
      required:
      - days
...
    FieldProblem:
      type: object
      properties:
        code:
          type: string
          example: min
        field:
          type: string
          example: reps.max
        message:
          type: string
      required:
      - code
      - field
      - message
...
    Problem:
      type: object
      description: Erro no formato RFC 9457 (application/problem+json)
      properties:
        code:
          type: string
          description: Código estável (kebab-case) que o app usa para escolher a mensagem
          example: resource-not-found
        detail:
          type: string
          example: Recurso não encontrado.
        errors:
          type: array
          description: Só em erros de validação (400)
          items:
            $ref: "#/components/schemas/FieldProblem"
        instance:
          type: string
          example: /v1/me
        status:
          type: integer
          format: int32
          example: 404
        title:
          type: string
          example: Recurso não encontrado
        traceId:
          type: string
          example: 4bf92f3577b34da6a3ce929d0e0e4736
        type:
          type: string
          example: https://api.moveup.com.br/problems/resource-not-found
      required:
      - code
      - detail
      - status
      - title
      - traceId
      - type
...
    PushDevice:
      type: object
      properties:
        platform:
          type: string
          enum:
          - ios
          - android
          minLength: 1
        token:
          type: string
          description: "ExponentPushToken[...]"
          maxLength: 200
          minLength: 0
      required:
      - platform
      - token
    PushDeviceRemoval:
      type: object
      properties:
        token:
          type: string
          maxLength: 200
          minLength: 0
      required:
      - token
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: ID token do Firebase Auth
      scheme: bearer
      type: http